--- a/rust/Cargo.toml
+++ b/rust/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "datason-rust"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "_datason_rust"
+crate-type = ["cdylib"]
+
+[dependencies]
+pyo3 = { version = "0.21", features = ["extension-module"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"